fix: Change back base.php to have the same behavior as before
Isolated OC in its own file instead so that index.php can do finer init.
