+++ lib/private/Route/Router.php
@@ -280,7 +280,6 @@ class Router implements IRouter {
 			$this->loadRoutes();
 		}
 
-
 		$this->eventLogger->start('route:url:match', 'Symfony url matcher call');
 		$matcher = new UrlMatcher($this->root, $this->context);
 		try {
